extended

+++ src/main/java/com/oop/OOP/repo/AdminRepository.java
@@ -3,5 +3,4 @@ package com.oop.OOP.repo;
 import com.oop.OOP.entities.Admin;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdminRepository extends JpaRepository<Admin, Integer> {
-}
+public interface AdminRepository extends JpaRepository<Admin, Long> {}
