+++ src/main/java/com/oop/OOP/repo/CandidateRepository.java
@@ -2,6 +2,6 @@ package com.oop.OOP.repo;
 import com.oop.OOP.entities.Candidate;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CandidateRepository extends JpaRepository<Candidate, Integer> {
-}
+public interface CandidateRepository extends JpaRepository<Candidate, Long> {}
+
